build-sys: enable static linking of externally-used programs
If downstreams want to ship these utils for use without systemd (eg, for containers
or other minimal systems), then they would have to have a copy of libshared. This
complicates packaging since everything would have to be upgraded in lockstep.

Not only that, but libshared links to most library dependencies of systemd, of which
a container that only requires the programs needs very few.